:art: Move serviço geracao de notificacoes para o modelo
Move o serviço que gera as notificações de vigência de contrato para
o modelo de notificação de vigência de contrato.
