+++ sme_coad_apps/contratos/tests/test_notificacao_vigencia_contrato_endpoints.py
@@ -12,3 +12,13 @@ def test_url_unauthorized(client):
 def test_url_authorized(authencticated_client):
     response = authencticated_client.get('/gera-notificacoes-vigencia-contratos/')
     assert response.status_code == status.HTTP_202_ACCEPTED
+
+
+def test_url_minhas_notificacoes_unauthorized(client):
+    response = client.get('/minhas-notificacoes-vigencia-contratos/')
+    assert response.status_code == status.HTTP_401_UNAUTHORIZED
+
+
+def test_url_minhas_notificacoes_authorized(authencticated_client):
+    response = authencticated_client.get('/minhas-notificacoes-vigencia-contratos/')
+    assert response.status_code == status.HTTP_200_OK
